edit PlayerController.cs and GunController.cs
