+++ Assembly-CSharp.csproj
@@ -54,9 +54,8 @@
      <Compile Include="Assets\Import Assets\TestLevel\scripts\explosive.cs" />
      <Compile Include="Assets\Import Assets\TestLevel\scripts\playrandomsound.cs" />
      <Compile Include="Assets\Import Assets\TestLevel\scripts\waitAndDestroy.cs" />
-     <Compile Include="Assets\Scripts\Bullet.cs" />
+     <Compile Include="Assets\Scripts\GunController.cs" />
      <Compile Include="Assets\Scripts\PlayerController.cs" />
-     <Compile Include="Assets\Scripts\test.cs" />
      <None Include="Assets\Import Assets\TestLevel\environment\skyshader1.0\mysky.shader" />
      <None Include="Assets\Import Assets\TestLevel\environment\customTrees\Materials\TreeCreatorLeavesFastOptimized.shader" />
      <None Include="Assets\Standard Assets\CrossPlatformInput\CrossPlatformInputGuidelines.txt" />
