--- Assets/Scripts/test.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 243469042e13a4427a334517329d8c41
-timeCreated: 1488107321
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
